requirements.txt
discord-py-interactions==5.6.0
openpyxl==3.1.2
pytz